+++ cprior/_lib/include/cprior.h
@@ -2,7 +2,7 @@
 extern "C" {
 #endif
 
-double cpp_beta_cprior(int a0, int b0, int a1, int b1);
+double cpp_beta_cprior(double a0, double b0, double a1, double b1);
 
 #ifdef __cplusplus
 }
