+++ cprior/_lib/include/cprior.hpp
@@ -1,6 +1,6 @@
 #ifndef CPRIOR_H
 #define CPRIOR_H
 
-double beta_cprior(int a0, int b0, int a1, int b1);
+double beta_cprior(double a0, double b0, double a1, double b1);
 
 #endif /* CPRIOR_H */
