+++ cprior/_lib/src/capi.cpp
@@ -2,7 +2,7 @@
 #include "cprior.h"
 
 
-double cpp_beta_cprior(int a0, int b0, int a1, int b1)
+double cpp_beta_cprior(double a0, double b0, double a1, double b1)
 {
   return beta_cprior(a0, b0, a1, b1);
 }
